requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.12.2
billboard.py==7.0.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
